pyproject.toml
[tool.poetry]
name = "fastapi-blog"
version = "0.1.0"
description = ""
authors = ["yetian29 <dathien000@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pydantic-settings = "^2.6.0"
motor = "^3.6.0"
punq = "^0.7.0"
fastapi = "^0.115.4"
uvicorn = "^0.32.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.7.1"
polyfactory = "^2.17.0"
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
    "W191",  # indentation contains tabs,
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["test"]